Upgrades finish (2/2 step complete)

+++ Drill Game/Assets/Scripts/Upgrades/LevelableEntity.cs
@@ -5,9 +5,9 @@ namespace Upgrades
     public abstract class LevelableEntity : MonoBehaviour
     {
         public int Level { get; private set; } = 1;
-        public int StartValue { get; protected set; }
-
-        public virtual void Upgrade(int value) 
+        public float StartValue { get; protected set; }
+        
+        public virtual void Upgrade(float value) 
         {
             Level++;
         }
